fixed bug with not showing all playlist tracks; refactored Tracks.js
